+++ src/FontStashSharp/Bounds.cs
@@ -2,6 +2,6 @@
 {
 	struct Bounds
 	{
-		public float X, Y, X2, Y2;
+		public double X, Y, X2, Y2;
 	}
 }
