+++ src/FontStashSharp/FontGlyphSquad.cs
@@ -5,13 +5,13 @@ namespace FontStashSharp
 	[StructLayout(LayoutKind.Sequential)]
 	struct FontGlyphSquad
 	{
-		public float X0;
-		public float Y0;
-		public float S0;
-		public float T0;
-		public float X1;
-		public float Y1;
-		public float S1;
-		public float T1;
+		public double X0;
+		public double Y0;
+		public double S0;
+		public double T0;
+		public double X1;
+		public double Y1;
+		public double S1;
+		public double T1;
 	}
 }
