+++ src/SpriteBatchExtensions.cs
@@ -6,12 +6,12 @@ namespace SpriteFontPlus
 {
 	public static class SpriteBatchExtensions
 	{
-		public static float DrawString(this SpriteBatch batch, DynamicSpriteFont font, string text, Vector2 pos, Color color)
+		public static double DrawString(this SpriteBatch batch, DynamicSpriteFont font, string text, Vector2 pos, Color color)
 		{
 			return font.DrawString(batch, text, pos, color);
 		}
 
-		public static float DrawString(this SpriteBatch batch, DynamicSpriteFont font, string text, Vector2 pos, Color color, Vector2 scale)
+		public static double DrawString(this SpriteBatch batch, DynamicSpriteFont font, string text, Vector2 pos, Color color, Vector2 scale)
 		{
 			return font.DrawString(batch, text, pos, color, scale);
 		}
